ML/CNN/pyproject.toml
[tool.poetry]
name = "cnn"
version = "0.1.0"
description = ""
authors = ["Chu Minseok <mchu7797@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = {version = "^2.4.1+cu124", source = "torch"}
torchvision = {version = "^0.19.1+cu124", source = "torch"}
torchaudio = {version = "^2.4.1+cu124", source = "torch"}
matplotlib = "^3.9.2"
pillow = "^10.4.0"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu124"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
